Access control

+++ session/crud/index.php
@@ -3,7 +3,17 @@ require_once "inc/functions.php";
 $info = '';
 $task = $_GET['task'] ?? 'report';
 $error = $_GET['error'] ?? '0';
+if('edit'==$task){
+    if(!hasLogIn()){
+        header("location:../index.php?task=report");
+
+    }
+}
 if ('delete' == $task) {
+    if(!isAdmin()){
+        header("location:../index.php?task=report");
+        return;
+    }
     $id = filter_input(INPUT_GET, 'id', FILTER_SANITIZE_STRING);
     if ($id > 0) {
         deleteStudent($id);
@@ -12,6 +22,10 @@ if ('delete' == $task) {
 
 }
 if ('seed' == $task) {
+    if (!isAdmin()) {
+        header("location:../index.php?task=report");
+        return;
+    }
     seed();
     $info = 'Seeding  is Complete';
 }
